Moved the tables and changed the intial values
